WIP move net.imglib2.ui classes
BufferedImageOverlayRenderer -> bdv.viewer.render.awt
BufferedImageRenderResult    -> bdv.viewer.render.awt
GuiUtil (as AWTUtils)        -> bdv.util
