+++ src/main/java/bdv/viewer/render/awt/BufferedImageRenderResult.java
@@ -1,5 +1,7 @@
-package bdv.viewer.render;
+package bdv.viewer.render.awt;
 
+import bdv.util.AWTUtils;
+import bdv.viewer.render.RenderResult;
 import java.awt.geom.AffineTransform;
 import java.awt.image.AffineTransformOp;
 import java.awt.image.BufferedImage;
@@ -8,7 +10,6 @@ import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.display.screenimage.awt.ARGBScreenImage;
 import net.imglib2.realtransform.AffineTransform3D;
 import net.imglib2.type.numeric.ARGBType;
-import net.imglib2.ui.util.GuiUtil;
 
 public class BufferedImageRenderResult implements RenderResult
 {
@@ -17,12 +18,8 @@ public class BufferedImageRenderResult implements RenderResult
 	private int width;
 	private int height;
 
-	// TODO: rename
-	//  renderData?
 	private int[] data = new int[ 0 ];
 
-	// TODO: rename
-	//  renderImage, img, image?
 	private ARGBScreenImage screenImage;
 
 	private BufferedImage bufferedImage;
@@ -42,7 +39,7 @@ public class BufferedImageRenderResult implements RenderResult
 			data = new int[ width * height ];
 
 		screenImage = new ARGBScreenImage( width, height, data );
-		bufferedImage = GuiUtil.getBufferedImage( screenImage );;
+		bufferedImage = AWTUtils.getBufferedImage( screenImage );;
 	}
 
 	@Override
